Get a working, first-pass core.py w/ Tomayto class
This is gross atm; I'm making a Tomayto class, instancing it immediately
after to a tomayto instance, and in the class, referring not just to the
instance directly, but to the nested modules as I'm using them on my
computer currently. Ugh. I have to figure out how to marshall things so
I don't have python("tomayto.core.tomayto.etc...") calls hardcoded into
the methods of the class. I really wish nameCommand's sourceType flag
was working, so I could use Python. It's endlessly harder to have to
stringify everything through MEL and back to Python :(