Add first, fledgling state machine for posterity
This is what I've wanted to see for years now! I can press 3 different
keys to make a ball, cube, or cylinder appear in Maya, and let go to
watch them disappear. Caveat: I have to press and release one key at a
time, because it's a really simplistic system that took a few minutes to
put together, just to see something happen. Exciting, though!