Add first pass play with arpeggiation
This is such a hard problem. I want the following:

    - mapping for keys
    - mapping for chords
    - arpeggiation, for lack of n-key rollover
    - mapping for chord/arpeggio release

Having mappings for release is a really fun idea, giving temporary-on
push button functionality, but it complects everything.

At first I thought I'd add chords by recording keypresses in a dict,
with values of when they were pressed in ms (time.time()), removing
keys on new keypresses if they were too old. The weird thing here is
that keys enter and fall out of a chord, which means it becomes harder
and harder to arpeggiate, the longer they get.

Then I thought I'd wait 0.1s after each key press, and use a kind of
keep-alive method of holding onto the chain of arpeggiated keys. This is
nice, but what terminates it? How does the last keypress ever trigger,
if we have to wait 0.1s after it's pressed? We could sleep(), but then
you'd never be able to read the next keypress, I think, and I just don't
like locking up the program, waiting.

I found the threading module, which allows waiting, and the timer could
just be pushed forward, or a myriad other things, handling a callback to
terminate, so I'll probably solve something along those lines. It
doesn't answer other questions, though.

One problem is what to do if keys are released during a chord, or if
more are pressed immediately after. Should the instant we have a full
chord, which isn't a prefix of another chord, trigger it? What do we do
with any keys that came in after? I started to feel like we needed a
buffer to track keys pressed, with times, and then have something that's
always actively working on the buffer, which I didn't like at all.

I think, now, that I need chords to work their own way. Any time you
press a key, there needs to be a threading module pause for 0.1s, with a
callback, to see if you're starting a chord. If you're not, the callback
tells something to go ahead with that key press. If another key comes in
before the callback, it locks some property, and then the callback
simply unlocks it and gives up. If the callback encounters the unlocked
property, then it knows another key hasn't come in since it unlocked it,
and it looks up the chord, firing it off, if it exists. This simplifies
things greatly; once you'ge begun chording or arpeggiating, by hitting a
second key without 0.1s of the first, you're only doing that. You'll
never be able to fire off a non-chord/arpeggio until you let that chord
end. I don't have to check if the first key, or the first two, etc., are
part of any chords; I just wait until you're done chording, and do a
lookup for the keys that made it in. This means if you set up a chord
for CAT, and someone chords CATS, it won't take. It's the wrong chord.

Then I need to figure out if it makes sense to allow chorded release
mappings. Do they fire once all keys are released? What if you release
and repress a key, while releasing others? Do they hang on? What if you
press a new key before releasing those? I think maybe releasing doesn't
make as much sense, though I'd like to solve it. It's a neat extra
feature. If it's too much, though, it's too much.