Change default callback handler method name
I thought "getch" made sense before; it's the way you get the character
input by the user in other languages, and this key-based state machine
idea, which I've implemented in various forms in past years, has used a
getch. However, it just feels confusing in this context. It's really a
default handler method, currently only manually hooked up after
instantiating the class, and only for use in testing if the getting of
pressed keys is working. It's really not a getch. The many nameCommands
and hotkeys, defined elsewhere in the class, share that role.