Add working core of Tomayto class
These methods form the first working version of the Tomayto class. It's
a bit cumbersome to create states for, and I don't know how many hidden
gotchas lurk, and it was a lot of sculpting of the idea over the course
of a day, rather than clean implementation, but it works.

Now the task will be to start populating the state machine, using it,
seeing what works, and looking for better abstractions, a more
declarative methodology, better ergonomics, et al.