Add run handling to eventHandler
This is just like popping, in that it passes an instance-local method to
run, but uses the "RUN" tag instead, which skips the changes to the
stateStack. If pushing and popping are like calling a function, and
receiving a result from it, running is like performing an action, in the
Haskell sense. It's a [typically] IO operation w/ no arguments, nor
return value.