Add a fledgling help... thing
This is an example of where a centralized handler is cool. I can go
right to the source and interrupt the keyhandling to print out info on
the current state, and its keymap events. This will also get very messy,
very fast, and needs some kind of pleasant system around it.