Allow passing args to pushState
Previously, you could pass the state name. Now you can pass the state
name, or, optionally, a tuple of the state name and a list of args. The
args are passed to the state class on instantiation, and they're passed
with *, so they become regular, non-list args to your state class' init
function. You can package them up in a list again, if you want, by using
another * on the appropriate/final state class init method arg.

The impetus for this feature was realizing I have many Maya window
layouts that are all similar, and might best be handled by a single
state that simply differentiates between them for their minor
differences, handling everything else the same. I'm reconsidering the
idea for this particular example, but I think it's a nice, and maybe
highly useful design decision, regardless.