Add [slightly hackish] Backspace handling hack
Backspace is not available to bind as a hotkey in Maya. You can see that
something is bound to Ctrl+Backspace in the visual hotkey editor, but
there's just no option to bind it, nor the unmodified version, nor any
other combo of modifiers + Backspace yourself; it tells you flat out
that you tried to bind to Backspace (plus any modifiers), and that it
isn't allowed. That said, I noticed in Maya that pressing Backspace, and
also Delete, printed out "doDelete;" in the output window. I created
the following global proc, and learned that I could easily override
Maya's handler:

    global proc doDelete ()
    {
        print "MUAHAHAhahahha";
    }

I also wanted to be able to restore Maya's handling, so there's some
complicated, probably wrong logic in here to find and store the path to
the script that creates doDelete in an optionVar, and then source that
script again to unhack Backspace.