Get backspace hack working on Windows
I'm on Maya 2017 on Linux at home, and Maya 2018 on Windows at work
(where I also tested 2016). At home, I get the familiar-to-me response
from whereIs: "Mel Procedure found in: etc..." At work, for some reason,
in both 2016 and 2018 I get the unfamiliar "Script found in: etc..." So,
I tried to make it work reasonably for both. I'm sure the logic could be
better here, but it worked at work, and found the path. Now to try it
again at home, and see if it still finds the path there.

Maya 2018 on Windows at work:
    C:/Program Files/Autodesk/Maya2018/scripts/others/doDelete.mel

Maya 2017 on Linux at home (maya is a symlink to maya2018):
    /usr/autodesk/maya/scripts/others/doDelete.mel