Add in a test window version of VimLine
This is really starting to be fun. There are plenty of issues, like the
cursor position in various states, and it's some sloppy, all over the
place, type-free, free-wheeling, cowboy style programming, but I don't
see a way to make it much better currently. I'm hoping as I keep playing
and building, that I'll see opportunities to greatly simplify and
condense the concepts into something with a bit more rigor.

I feel encumbered by state from all directions, which I guess feels
appropriate, given that I'm making a state machine of keypresses, which
is twiddling state in Maya, Maya UIs, and Python. It's all so stateful,
and state is a bear to work with.