Add tool selection state from start state
I don't know why this is necessary, but it's nice to, e.g., turn off
seeing the manipulators when selecting things (currently only meshes),
by selecting the select tool. This changes the defaults every Maya user
has known for about 2 decades, namely w for move, e for rotate, and r
for scale, making them, more sensibly m, r, and s, respectively, and
grouping them under a t (for tool) prefix, but I recognize it will
likely screw with our muscle memory when switching back to default Maya.

This is also the moment I've realized I need an easy to way to appeal to
the main instance (mainInst) for things like auto-pop on pressing a key
that doesn't exist in the map, so ts gives you the scale tool, but tj
should just pop back to the start state, as there is no j tool.