Allow Enter to exit vimline, and its insert mode
Because there's only one line, Enter can't mean to insert a carriage
return, nor newline, and it feels in the test window like filling in one
line of information anyway, so it feels natural to escape by hitting
Enter when done. I tried this out, and it felt good, and natural.