Generalize mesh selection to any transform
This is a simpler version of what I ultimately want, which is full
filtering and boolean operations over selections. Also of note, I had to
hack the args in some onPopTo methods to take *args, instead of just an
arg; I hope this gets cleared up in future revisions of the state stack
system; this is the second time I've hit silent failures on popping.