Remove all stateMaps
It turns out that stateMaps are redundant. I recently made the change to
pass the module level map of states into the state stack - which allowed
me to reference states in other files by their maps, allowing me to
split states up over multiple files - along with the name of the state
(key in the map), which pointed at the actual state class reference, but
it's easier, and equivalent, to just pass the ref to the state,
qualified by its module, where necessary, and forget the maps and keys.