Fix push and push event tests
I had one class with an onEnter callback that created a locator as a
witness that the callback was indeed called, but which was being tested
both through a push event match, and a direct pushing. Because the Maya
scene is not refreshed (and doing so would cause its own problems with
the maya nose @attr filter), one of the two tests was running in a world
where the locator already existed. I had to duplicate all the pathways,
and come up with longer, crazier names to separate them out.