Revamp all core tests to avoid need for maya.cmds
I realized that instead of using Maya as the global space, and locators
as witnesses to methods being run, proving they'd been called, either
explicitly, or through events triggering, I could simply use the much
maligned global variables, and do away entirely with the need to run
things through the much slower Maya testing pathway. Maya may make a
return to this file's tests, but for now, it's nice not to need it.