Change push event key from 'u' to 'p'
Super trivial change, but I made it 'u' earlier, because push and pop
both start with 'p', so I decided to use 'u' and 'o', their second
letters, to make it easy to follow what the events were for. Now I'm
realizing that I'll never have a pop event in the start state, so the
push event can just use 'p' after all.