Switch transform select tweaks away from recursion
Thanks to Elbenoa in the Twitch chat for the idea to simply run a method
to update these values, instead of what I thought was the simple,
obvious solution of recursing into the same state with new arg values.
This also now properly updates the annotations, when needed. Thanks
again to Ben for noticing ~5 errors in my code as I made the changes.