Remove label texts from SelectionList
Too fussy, looked ugly, hard to manage - I've decided to just make it
easy to change the text in the text controls, and managing what they say
is up to you. I'm pulling back from the SelectionList having a lot of
powers, and just making it a display target. It simplifies everything.