Begin redesigning scrolling text list
This throws out the unwieldy scrollLayout, in favor of a new (to me)
layout, the flowLayout. Scrolling is not yet working, but the new method
is based on some playings around I did in Maya, where instead of
actually scrolling, you'll be using an int slider to choose an index in
the entries list, and then everything before that will simply be hidden.
This solves a bunch of my wishes in one fell swoop, including keeping
the topmost entry always perfectly registered at the top, and all other
entries (when all heights match) registered to discrete positions in the
scroll area. It also makes it easy to scoll variable-height entries to
the top, exactly, every time. It also means I can scroll past the bottom
of the list, which I wanted, even if it's a little weird. The selection
keys are my attempt to go from easiest to press to harder ones, but it
doesn't really make sense to do that anyway, as any item in the list can
be the most important.