Handle scrolling by mouse
This leaves the key selection column alone, except when the entries list
is scrolled up past it, when it hides any keys that no longer match up
with entries. This has no effect on the as yet unimplemented actual key
press handling. This just hides entries, but doesn't disable anything.