Ensure non-duplicated values during init
This is something that's popped into my head over and over as I've
pondered this UI, and I just ran into a reason to require it, finally.
Scrolling the list suddenly started leaving remnants behind, and always
the same ones, so as you scrolled down, more and more of the topmost
visible elements would freeze in place, and unfreeze as you scrolled
back up. It was hard to track down, but finally clicked; now that we're
using a dictionary lookup, keys have stomped over by later dupes.