Add filter/sort ability via passed-in functions
The values property holds the passed-in values for the list, and entries stores
the actual UI elements in a dict, along with their selected states. Now, the
values are filtered and sorted just before UI population, and the entries are
looked up from there. This does not yet clear out the ui keys of later filtered
entries.