Add scrolling to the top and bottom
I had to use 6 and 4 (with ctrl) as stand-ins for ^ and $, Vim's motions
for jumping to the beginning and end of a line, respectively. Normally,
to jump to the beginning and end of the buffer, you'd use gg and G,
respectively, but all lower and upper case letters are for selection
toggling, I've decided, and I don't want to deal with a 2-key event yet.