Add toggleVisible method
I was trying to make this work for visible, or all filtered, but
realized that my way of selecting values was through toggling, and it
currently only allows toggling by associated key, and only visible
entries have keys, and even visible ones off the bottom of the
flowLayout, with associated keys, are skipped, as a design decision, as
it would mean you could accidentally toggle selection on things you
can't see. So, I simplified; you can only toggle selection on all
visible.