Add getVimlineParts
This greatly simplifies asking for the left, cursor, and right parts of
the vimline line of text, and is only really needed as things gets weird
over at the eol end of things, e.g. the cursor can't go off the right
end in normal mode, as it can in other editors, and in insert mode.