Add a showable/hideable top pane for user controls
E.g. for a search bar. It's a paneLayout, with some vimline-named text
controls atm, but I think those will come out, and the idea will be that
clients of a SelectionList will be given a reference to that pane, and
it'll be up to them to put whatever they want into it, and use the hide
and show abilities of the pane as they see fit.