Add backspace/delete hack to make them usable
This existed on an old branch, which got merged in with strategy "ours",
and so didn't become part of the codebase. I learned, in reviving it,
that my backspace hack also hacks delete; they both call the same MEL
function (doDelete), and there's no way to discern which was pressed. I
also can't detect modifiers, or press vs. release, so we can only detect
that either key was pressed, modifiers or not, but at least it'll be
easy to backspace when typing through the Vimline module now.